# DataInput: Unroll loop in readVInt and readVLong